% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobieraczek2.R
\docType{data}
\name{args}
\alias{args}
\alias{pobieraczek}
\title{Pobieranie informacji o filmach z bazy IMDb wersja 2.}
\format{\preformatted{ chr(0) 
}}
\usage{
args
}
\arguments{
\item{from}{filmy od roku...}

\item{to}{filmy do roku... wlacznie.}

\item{min}{minimalna liczba linijek do pobrania}

\item{maksymalna}{liczba linijek do pobrania}
}
\value{
zapis do pliku(ow) w folderze ~/dane/glowne/rok.csv dane filmow z podanych lat.
}
\description{
Funkcja \code{pobieraczek2} pobiera wszystkie dane o filmach z podanych lat i zapisuje do
ramki danych w folderze ~/dane/glowne/rok.csv.
}
\keyword{datasets}

