% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_movie_url_year.R
\name{download_movie_url_year}
\alias{download_movie_url_year}
\title{Pobieranie linkow do filmow na IMDb}
\usage{
download_movie_url_year(from, to = from)
}
\arguments{
\item{from}{filmy od roku...}

\item{to}{filmy do roku... wlacznie.}
}
\value{
zapis do pliku(ow) w folderze ~/movies_link ramki danych postaci |rok|link|.
}
\description{
Funkcja \code{download_movie_url_year} pobiera wszystkie linki do filmow z bazy IMDb. Wynikiem
jest dwukolumnowa ramka danych postaci | year | link |. Ramka zapisywana jest do roznych plikow
z rozdzieleniem na lata.
}

