% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/from_full_cast_n_crew.R
\name{from_full_cast_n_crew}
\alias{from_full_cast_n_crew}
\title{Pobieranie tabel z podstrony obsady portalu IMDb.}
\usage{
from_full_cast_n_crew(link, sep = ",")
}
\arguments{
\item{link}{link do strony glownej filmu na IMDb'ie, \code{typu character}}

\item{sep}{znak rozdzielania wielu informacji w jednej komorce danych}
}
\value{
jednowierszowa ramka danych o nazwach kolumn: "DirectedBy"|"Cast"|"Writing"|"ProducedBy"|"MusicBy"|"CinematographyBy"
}
\description{
Funkcja \code{from_full_cast_n_crew} zwraca jednowierszowa ramke danych z informacja o rezyserze,
obsadzie oraz autorach scenariusza, muzyki i zdjec. Ewentualnych kilka nazwisk (jak np. przy obsadzie)
rozdzielanych jest separatorem \code{sep}.
}

