% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/from_main_page.R
\name{from_main_page}
\alias{from_main_page}
\title{Pobieranie informacji ze strony glownej danego filmu na IMDb.}
\usage{
from_main_page(link, sep = ",")
}
\arguments{
\item{link}{link do strony glownej filmu na IMDb'ie, \code{typu character}}

\item{sep}{znak rozdzielania wielu informacji w jednej komorce danych}
}
\value{
jednowierszowa ramka danych o nazwach kolumn: "title"|"year"|"duration"|"genres"|"rating"|"votes"
}
\description{
Funkcja \code{from_main_page} zwraca jednowierszowa ramke danych z informacja o tytule, roku powstania,
czasu trwania, gatunku i ocenie filmu oraz liczbie oddanych nan glosow. Dane pobiera uzywajac
roznych typow nodes po sparsowaniu strony glownej. Ewentualnych kilka nazw (jak np. przy gatunku)
rozdzielanych jest separatorem \code{sep}.
}

