% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/from_page_source.R
\name{from_page_source}
\alias{from_page_source}
\title{Pobieranie informacji z kodu zrodlowego strony glownej filmu w bazie IMDb.}
\usage{
from_page_source(link, sep = ",")
}
\arguments{
\item{link}{link do strony glownej filmu na IMDb'ie, \code{typu character}}

\item{sep}{znak rozdzielania wielu informacji w jednej komorce danych}
}
\value{
jednowierszowa ramka danych o nazwach kolumn: "production_countries"|"language"|"color"
}
\description{
Funkcja \code{from_page_source} zwraca jednowierszowa ramke danych z informacja o krajach produkcji,
jezyku danego filmu oraz wersji kolorystycznej. Dane wyszukiwane sa bezposrednio z kodu
zrodlowego danej strony. Ewentualnych kilka nazw (jak np. przy gatunku) rozdzielanych jest
separatorem \code{sep}.
}

