% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keywords.R
\name{keywords}
\alias{keywords}
\title{Pobieranie slow kluczowych o filmie.}
\usage{
keywords(link, sep = ",")
}
\arguments{
\item{link}{link do strony glownej filmu na IMDb'ie, \code{typu character}}

\item{sep}{znak rozdzielania wielu informacji w jednej komorce danych}
}
\value{
jednowierszowa ramka danych o nazwach kolumn: "title"|"year"|"duration"|"genres"|"rating"|"votes"
}
\description{
Funkcja \code{keywords} pobiera i zapisuje do wektora kluczowe slowa dotyczace filmu.
}

