% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_csv.R
\name{merge_csv}
\alias{merge_csv}
\title{Sklejanie plikow}
\usage{
merge_csv(folder)
}
\arguments{
\item{folder}{nazwa folderu z plikami .csv}
}
\value{
ramka danych
}
\description{
Funkcja \code{merge_csv} skleja do jednej ramki danych wszystkie pliki .csv ze wskazanego folderu.
Sluzy do laczenia w jedna tabele pobranych linkow do filmow.
}

