% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pobieraczek.R
\name{pobieraczek}
\alias{pobieraczek}
\title{Pobieranie informacji o filmach z bazy IMDb.}
\usage{
pobieraczek(from, to = from)
}
\arguments{
\item{from}{filmy od roku...}

\item{to}{filmy do roku... wlacznie.}
}
\value{
zapis do pliku(ow) w folderze ~/dane/glowne/rok.csv dane filmow z podanych lat.
}
\description{
Funkcja \code{pobieraczek} pobiera wszystkie dane o filmach z podanych lat i zapisuje do
ramki danych w folderze ~/dane/glowne/rok.csv.
}

