% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rating_stats.R
\name{rating_stats}
\alias{rating_stats}
\title{Rozklad glosow}
\usage{
rating_stats(link)
}
\arguments{
\item{link}{link do strony glownej filmu na IMDb'ie, \code{typu character}}
}
\value{
trzykolumnowa ramka danych o nazwach kolumn: "Votes"|"Percentage"|"Rating"
}
\description{
Funkcja \code{rating_stats} zwraca trzykolumnowa tabelke z rozkladem poszczegolnych glosow.
}

