% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save_table_rating.R
\name{save_table_rating}
\alias{save_table_rating}
\title{Pobieranie i zapis do pliku statystyk o filmie z podanego roku}
\usage{
save_table_rating(from, to)
}
\arguments{
\item{from}{filmy od roku...}

\item{to}{filmy do roku... wlacznie.}
}
\value{
zapis do pliku(ow) w folderze ~/dane/statystyki/rok.csv dane filmow z podanych lat.
}
\description{
Funkcja \code{save_table_rating} pobiera i zapisuje do ramki danych statystyki
wszystkich filmow z podanych lat.
}

