% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stats_to_one_row.R
\name{stats_to_one_row}
\alias{stats_to_one_row}
\title{Statystyki jednego filmu}
\usage{
stats_to_one_row(link)
}
\arguments{
\item{link}{link do strony glownej filmu na IMDb'ie, \code{typu character}}
}
\value{
jeden wiersz formatu ramki danych ze statystykami dotyczacymi filmu.
}
\description{
Funkcja ulatwia pobranie i zapis wszystkich statystyk o danym filmie do jednego pliku.
}

