% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table_rating.R
\name{table_rating}
\alias{table_rating}
\title{Statystyki wielu filmow}
\usage{
table_rating(links, name_of_table)
}
\arguments{
\item{links}{wektor linkow do stron glownych filmow na IMDb.}

\item{name_of_table}{nazwa pliku wyjsciowego}
}
\value{
ramka danych za statystykami filmow z linkow podanych parametrem links.
}
\description{
Funkcja ulatwia pobranie i zapis (w getwd()) wszystkich statystyk o wielu filmach do jednego pliku.
}

