% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/user_rating_stats.R
\name{user_rating_stats}
\alias{user_rating_stats}
\title{Rozklad glosow w grupach plciowo-wiekowych}
\usage{
user_rating_stats(link)
}
\arguments{
\item{link}{link do strony glownej filmu na IMDb'ie, \code{typu character}}
}
\value{
trzykolumnowa ramka danych o nazwach kolumn: "Who"|"Votes"|"Average"
}
\description{
Funkcja \code{user_rating_stats} zwraca trzykolumnowa tabelke z rozkladem glosow w poszczegolnych grupach
plciowo-wiekowych.
}

